Fix handling of binary, timestamp, and boolean types
hexspace has commented out code for timestamps and treats binary values
as strinsg.  Unfortunately, the hexspace code type conversion code is
in the middle of a large function named process_results, with no
simple way to override the parts we need.

Subclass Hexspace::Client, copy the process_results method into it,
and modify it to support timestamps and binary values.

In addition, use the correct Spark types for File, and literalize
files using the Spark to_binary function with a base64 format.

Override literal false/true to use the Spark expected values.