Make type_test.rb pass
Avoid the use of UPDATE statements, which Spark does not support
and are not yet emulated.

Remove the guards since these tests only target Spark.

Remove support for the time type, since Spark does not support
time type, only timestamp.

Spark does not support INSERT with DEFAULT VALUES, or DEFAULT
as a value when inserting, so it is currently not possible to
insert a row with all default values in Spark.  Update the boolean
test to test for a column default by using two columns.