Add UPDATE emulation
Emulating UPDATE is a little more complex than DELETE, but the
idea is similar.  Copy all rows not being updated into a new
table (as delete).  Then insert into that table all updated
values from the current dataset.  Then drop the current table
and rename the new table to the current table.

There are a lot of tests that use UPDATE, this should be simpler
than modifying the tests.

Fix DELETE emulation to work correctly on datasets with explicit
selections. Also, fix the return value to return the number of
rows, which is determined by running a count first.