Add support for creating views without source dataset/sql
Spark supports views with only options, if the options specify an
appropriate data source.  This supports the following additional
options:

:using :: USING value, should be a string, used literally

:options :: OPTIONS value, should be a hash with keys and values
            converted to strings and then literalized

:temp :: Specifies creating a TEMPORARY view

The committed test depends on the examples/src/main/resources/users.parquet
file, which is shipped with Spark.  The file could potentially change
in future versions and cause the test to break.  However, it's worth
having an actual test.

In addition to a real integration test, add a test for specific SQL
for this feature, using a mock database that supports Spark syntax.
This also tests the mock Spark adapter works.