Handle forward references in CTEs in Dataset#with
When calling Dataset#with, if there are already CTEs for the
dataset, see if an existing dataset has a reference to the CTE
being inserted, and if so, insert the new CTE before the CTE
that references it.

This uses a Sequel::ASTTransformer subclass to try to find all
references.  This should handle common scenarios, but I would
guess there are still ways it can miss references.  Certainly
one way it can miss references is if mixing normal references
with unquoted literal string references, since this relies
on an matching literalized identifiers.