Make Spark as supporting CTEs in subqueries
Spark appears to handle CTEs in subqueries fine, and this fixes
a no method [] on NilClass that occurred previously when calling
Dataset#with on a dataset that already used a subquery with a
CTE, because the hoisted CTE would set :with without
:with_references.

For CTEs in compounds (e.g. UNION), Spark does not support those
directly in syntax, but you can wrap the compound dataset in a
subquery, and then it is supported as in any other subquery.